+++ src/main/java/bjs2_57242/Coordinates.java
@@ -1,4 +1,4 @@
-package school.faang;
+package bjs2_57242;
 
 public record Coordinates(int x, int y) {
 }
