+++ src/main/java/bjs2_57242/FlipDirection.java
@@ -1,4 +1,4 @@
-package school.faang;
+package bjs2_57242;
 
 public enum FlipDirection {
     VERTICAL,
