+++ src/main/java/bjs2_57242/MatrixTransformer.java
@@ -1,4 +1,4 @@
-package school.faang;
+package bjs2_57242;
 
 public interface MatrixTransformer {
     Coordinates transform(int x, int y, int rowCount, int columnCount);
