+++ src/test/java/bjs2_57242/MainTest.java
@@ -1,10 +1,8 @@
-package school.faang;
+package bjs2_57242;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class MainTest {
